+++ conftest.py
@@ -1,6 +1,5 @@
 import numpy as np
 import pytest
-from requests import RequestException
 
 from pyquil.api import (
     QVMConnection,
@@ -8,9 +7,7 @@ from pyquil.api import (
     Client,
     BenchmarkConnection,
 )
-from pyquil.api._errors import UnknownApiError
-from pyquil.api._abstract_compiler import QuilcNotRunning, QuilcVersionMismatch
-from pyquil.api._qvm import QVMNotRunning, QVMVersionMismatch
+from pyquil.api._qvm import QVMVersionMismatch
 from pyquil.device import Device
 from pyquil.gates import I
 from pyquil.paulis import sX
@@ -126,22 +123,17 @@ def qvm(client: Client):
         qvm = QVMConnection(client=client, random_seed=52)
         qvm.run(Program(I(0)), [])
         return qvm
-    except (RequestException, QVMNotRunning, UnknownApiError) as e:
-        return pytest.skip("This test requires QVM connection: {}".format(e))
     except QVMVersionMismatch as e:
         return pytest.skip("This test requires a different version of the QVM: {}".format(e))
+    except Exception as e:
+        return pytest.skip("This test requires QVM connection: {}".format(e))
 
 
 @pytest.fixture()
 def compiler(test_device, client: Client):
-    try:
-        compiler = QVMCompiler(device=test_device, client=client, timeout=1)
-        compiler.quil_to_native_quil(Program(I(0)))
-        return compiler
-    except (RequestException, QuilcNotRunning, UnknownApiError, TimeoutError) as e:
-        return pytest.skip("This test requires compiler connection: {}".format(e))
-    except QuilcVersionMismatch as e:
-        return pytest.skip("This test requires a different version of quilc: {}".format(e))
+    compiler = QVMCompiler(device=test_device, client=client, timeout=1)
+    compiler.quil_to_native_quil(Program(I(0)))
+    return compiler
 
 
 @pytest.fixture()
@@ -156,14 +148,9 @@ def client():
 
 @pytest.fixture(scope="session")
 def benchmarker(client: Client):
-    try:
-        bm = BenchmarkConnection(client=client, timeout=2)
-        bm.apply_clifford_to_pauli(Program(I(0)), sX(0))
-        return bm
-    except (RequestException, TimeoutError) as e:
-        return pytest.skip(
-            "This test requires a running local benchmarker endpoint (ie quilc): {}".format(e)
-        )
+    bm = BenchmarkConnection(client=client, timeout=2)
+    bm.apply_clifford_to_pauli(Program(I(0)), sX(0))
+    return bm
 
 
 def _str_to_bool(s):